fix: address code indexer review comments
